+++ src/pages/ReviewList.tsx
@@ -2,47 +2,22 @@ import KeywordStats from '@/components/KeywordStats';
 import ReviewSortingList from '@/components/ReviewSortingList';
 import ReviewStats from '@/components/ReviewStats';
 import { reviews } from '@/data/reviewData';
+import useMessageToParent from '@/hooks/useMessageToParent';
 import { ReviewType } from '@/types/RivewType';
 import { Margin } from '@/ui/margin/margin';
 import { Fonts } from '@/utils/GlobalFonts';
-import React, { useEffect } from 'react';
+import React, { useState } from 'react';
 import { styled } from 'styled-components';
 
 export default function ReviewList() {
-  const componentRef = React.useRef<HTMLDivElement>(null);
+  const { componentRef, setHeightChange } = useMessageToParent();
 
-  const [heightChange, setHeightChange] = React.useState(false);
-
-  const [rating, setRating] = React.useState(5.0);
+  const [rating, setRating] = useState(5.0);
 
   // 5,4,3,2,1점
-  const [scoreList, setScoreList] = React.useState([80, 60, 40, 20, 30]);
+  const [scoreList, setScoreList] = useState([80, 60, 40, 20, 30]);
 
-  const [reviewList, setReviewList] = React.useState<ReviewType[]>(reviews);
-
-  useEffect(() => {
-    const handleMessage = (e: MessageEvent) => {
-      if (!e.data.type) return;
-      if (e.data.type === 'loaded') sendHeightToParent();
-      console.log('부모로 부터 온 메세지:', e.data);
-    };
-
-    window.addEventListener('message', handleMessage);
-
-    return () => {
-      window.removeEventListener('message', handleMessage);
-    };
-  }, []);
-
-  const sendHeightToParent = () => {
-    if (!componentRef.current) return;
-    const message = componentRef.current?.offsetHeight + 30;
-    window.parent.postMessage({ type: 'height', message: message }, '*');
-  };
-
-  useEffect(() => {
-    sendHeightToParent();
-  }, [heightChange]);
+  const [reviewList, setReviewList] = useState<ReviewType[]>(reviews);
 
   return (
     <Container ref={componentRef}>
